requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiosignal==1.3.1
alpaca-py==0.8.1
alpaca-trade-api==3.0.0
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
deprecation==2.1.0
distlib==0.3.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.12.0
flatbuffers==23.5.8
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
msgpack==1.0.3
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.0
preshed==3.0.8
protobuf==3.20.3
pydantic==1.10.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
smart-open==6.3.0
srsly==2.4.6
sseclient-py==1.7.2
thinc==8.1.9
thread6==0.2.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
virtualenv-clone==0.5.7
wasabi==1.1.1
websocket-client==1.5.1
websockets==10.4
yarl==1.8.2
tensorflow==2.12.0
